refactor(common/lmlayer): reverts the parent/subclass bit, no longer needed
